+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.15.0] - 2025-10-16
+- Added hiding expired resource exceptions
+- Fixed an issue with filtering resource exceptions by `--tags_filters`
+
 
 ## [5.14.0] - 2025-10-03
 - Updated the ami-initialize.sh script